:truck: Rename Sets Management dropdown
