+++ src/views/DataExploration/components/SetsManagementDropdown/index.module.scss
